% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calcMeans.R
\docType{methods}
\name{calcMeans}
\alias{calcMeans}
\alias{calcMeans,SummarizedExperiment-method}
\title{Calculate cluster means}
\usage{
calcMeans(d_se)

\S4method{calcMeans}{SummarizedExperiment}(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in \code{\link{SummarizedExperiment}}
format, containing cluster labels as a column in the row meta-data (from
\code{\link{generateClusters}}). Column meta-data is assumed to contain a factor
\code{marker_class}.}

\item{...}{optional arguments.}
}
\value{
\code{d_means}: \code{\link{SummarizedExperiment}} object, where rows =
  clusters, columns = samples, sheets (\code{assays} slot) = markers. The
  \code{metadata} slot contains variables \code{id_type_markers} and
  \code{id_state_markers}, which can be accessed with
  \code{metadata(d_means)$id_type_markers} and
  \code{metadata(d_means)$id_state_markers}.
}
\description{
Calculate cluster means 
(mean expression for each cluster-sample-marker combination).
}
\details{
Calculate mean marker expression for each cluster and sample 
(i.e. means for each cluster-sample-marker combination).

The data object is assumed to contain a factor \code{marker_class} in the column
meta-data (see \code{\link{prepareData}}), which indicates the protein marker class for
each column of data (\code{"type"}, \code{"state"}, or \code{"none"}).

The cluster means are required for testing for differential states within cell
populations, and for plotting purposes.

Variables \code{id_type_markers} and \code{id_state_markers} are saved in the
\code{metadata} slot of the output object. These can be used to identify the 'cell
type' and 'cell state' markers in the list of \code{assays} in the output
\code{\link{SummarizedExperiment}} object, which is useful in later steps of the
'diffcyt' pipeline.

Results are returned as a new \code{\link{SummarizedExperiment}} object, where rows =
clusters, columns = samples, sheets (\code{assays} slot) = markers. Note that there is
a separate table of values (\code{assay}) for each marker. The \code{metadata} slot
also contains variables \code{id_type_markers} and \code{id_state_markers}, which can
be used to identify the sets of cell type and cell state markers in the list of
\code{assays}.
}
