% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotPerf.R
\docType{methods}
\name{plotPerf}
\alias{plotPerf}
\alias{plotPerf,COBRAPerformance-method}
\title{TPR vs. FDR curve}
\usage{
plotPerf(x, ...)

\S4method{plotPerf}{COBRAPerformance}(x, main = NULL, cols = NULL)
}
\arguments{
\item{x}{output from \code{iCOBRA::calculate_performance}}

\item{...}{optional arguments.}

\item{main}{character string to use as plot title.}

\item{cols}{named colors palette to use for methods. 
Defaults to \code{RColorBrewer::brewer.pal(11, "Spectral")},}
}
\description{
Plots a TPR vs. FDR curve from the performance output of \code{iCOBRA}.
}
\author{
Helena Lucia Crowell
}
