% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runDiffcyt.R
\docType{methods}
\name{runDiffcyt}
\alias{runDiffcyt}
\alias{runDiffcyt,daFrame-method}
\title{...}
\usage{
runDiffcyt(x, ...)

\S4method{runDiffcyt}{daFrame}(x, data, log = TRUE, k = NULL,
  method = c("limma", "LMM"), fun = c("mean", "median"))
}
\arguments{
\item{x}{a \code{daFrame}.}

\item{...}{optional arguments.}

\item{data}{character string specifying the data to use.
Should be one of \code{assayNames(x)}.}

\item{log}{logical. Should log2 be taken before computing means/medians.}

\item{k}{numeric or character string.
Specifies the clustering from which to retrieve cluster IDs.
Must be one of \code{colnames(cluster_codes(x))}.
Defaults to the 1st clustering available.}

\item{method}{a character string.}

\item{fun}{character string specifying the summary statistic to use.}
}
\description{
...
}
\author{
Helena Lucia Crowell
}
