% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runEdgeR.R
\docType{methods}
\name{runEdgeR}
\alias{runEdgeR}
\alias{runEdgeR,daFrame-method}
\title{...}
\usage{
runEdgeR(x, ...)

\S4method{runEdgeR}{daFrame}(x, k = NULL, method = c("rawCounts",
  "normCounts", "scaledCPM"))
}
\arguments{
\item{x}{a \code{daFrame}.}

\item{...}{optional arguments.}

\item{k}{numeric or character string. Specifies the clustering to use.}

\item{method}{a character string.}
}
\description{
...
}
\author{
Helena Lucia Crowell
}
