% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runFDA.R
\docType{methods}
\name{runFDA}
\alias{runFDA}
\alias{runFDA,daFrame-method}
\title{...}
\usage{
runFDA(x, ...)

\S4method{runFDA}{daFrame}(x, k = NULL, reso = 50, n_perm = 200)
}
\arguments{
\item{x}{a \code{daFrame}.}

\item{...}{optional arguments.}

\item{k}{numeric or character string. Specifies the clustering to use.}

\item{reso}{resolution.}

\item{n_perm}{numeric. Nb. of permutations to use in creating
the null distribution. Passed to \code{\link[fda]{tperm.fd}}.}
}
\description{
...
}
\author{
Helena Lucia Crowell
}
