% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/simDD2.R
\docType{methods}
\name{simDD2}
\alias{simDD2}
\alias{simDD2,SingleCellExperiment,numeric-method}
\title{\code{scDD} simulation}
\usage{
simDD2(x, y, ...)

\S4method{simDD2}{SingleCellExperiment,numeric}(x, y, n_cells = 100,
  n_genes = 5, n_samples = 4, n_clusters = 1, rand_samples = FALSE,
  rand_clusters = FALSE, perc_type = 0, perc_de = c(0, 0.25),
  cat_probs = rep(0.25, 4), sd_range = c(1, 3), mode_fc = c(2, 3, 4),
  seed = 1)
}
\arguments{
\item{x}{a \code{SingleCellExperiment}.}

\item{y}{numeric vector of gene indices to simulate from.}

\item{...}{optional arguments.}

\item{n_cells}{numeric of length 1, 2, or \code{n_clusters}.
Specifies the nb. of cells per cluster. 
If a single value is provided, clusters will be equally-sized.
If a range is provided and \code{rand_clusters = TRUE}, 
cluster sizes will be sampled from therein.
If \code{n_clusters} values are provided, these will be used.}

\item{n_genes}{numeric. Nb. of genes.}

\item{n_samples}{numeric. Nb. of samples per condition.}

\item{n_clusters}{numeric. Nb. of clusters.}

\item{rand_samples, rand_clusters}{logical. 
Should sample/cluster sized be randomized?}

\item{perc_type}{numeric. Percentage of genes to set as type genes.}

\item{perc_de}{numeric of length 1 or 2. 
Specifies the percentage of state genes to be DE per cluster.
If a range is provided, the percentage will be sampled from therein.}

\item{cat_probs}{numeric of length 4 specifying the probabilities 
of DE genes to be of category DE, DP, DM, and DB, respectively.}

\item{sd_range}{numeric vector of length to specifying
the range of SDs of FCs to sample from.}

\item{mode_fc}{numeric vector of possible FCs b/w modes for DP, DM, and DB.}

\item{seed}{numeric. Random seed for sampling & \code{scDD} simulation.}

\item{zeros}{method to generate zeros.}
}
\value{
a \code{daFrame}.
}
\description{
Complex \code{scDD} simulation with 
varying numbers and sizes of samples and clusters.
}
\examples{
sce <- get(data("scDatEx", package = "scDD"))
gs <- scDD:::findIndex(sce, condition = "condition")
simDD(x = sce, y = gs)

}
\author{
Helena Lucia Crowell
}
